LPS-75604 Ignore locking errors when checking in/out
This is very ugly, but the alternative (asking the backend SP
repository for the file and testing the checke out status) can be
quite expensive -- definitelly more than catching an exception. Doing
it this way, we only pay the price for checked out file entries (which
are usually the minority).