LPS-76595 Manually filter search results
Manually filter search results to avoid an error when the returned
documents are outside of the current site.

Note that this shouldn't happen anymore, as we stopped using the
Documents search source, and the Local Sharepoint source doesn't seem
to have this problem. I'm adding this as a safeguard anyway in case an
administrator modifies the default parameters of the search source.